refactor: PR 의견 반영
