+++ app/src/main/java/com/example/swap_it/data/model/CardData.kt
@@ -8,7 +8,7 @@ data class ShoppingCardData(
     val viewCount: String,
     val region: String,
     val time: String,
-    val price: String,
+    val price: Int,
     val title: String,
     val onClick: () -> Unit = {}
 )
@@ -26,7 +26,7 @@ data class SwapCardData(
     val viewCount: String,
     val region: String,
     val time: String,
-    val price: String,
+    val price: Int,
     val title: String,
     val onClick: () -> Unit = {}
 )
